Documented generate_hint functions description
The goal of this function is to take two "5-character string inputs", the first being the actual word (word to guess to win the game), second being the word guessed by the player before using a hint. The Function extracts a random character from the "actual word" and returns it according to its relative index position.